fix(slides): data distributions lecture 02
The slide 6 method for using percent_rank to sort a column of
values into quartiles does not work correctly in all cases. The
number of values assigned to each quartile should be approximately
equal, but this method does not enforce that requirement. Slide 6
has been revised to use the ntile function instead, which should
yield quartile boundaries that are equivalent to the quantile
function in type = 1 mode.

On slide 5, ntile is now listed instead of percent_rank.