fix(slides): inference and simulation 5
Update slide 7 to show how to compute the observed statistic for
the college applications dataset using infer instead of providing
the 9.7 value manually. The infer-computed mean will still come out
to 9.7, but this shows students a more programmatic and less
error-prone method for calculating the observed statistic that can
be used in their code.